+++ Serialization/FO-DICOM.Json/JsonDicomConverter.cs
@@ -229,7 +229,7 @@ namespace FellowOakDicom.Serialization
                     }
                     break;
                 case "LO":
-                    item = new DicomLongString(tag, (string[])data);
+                    item = new DicomLongString(tag, _jsonTextEncoding, (string[])data);
                     break;
                 case "LT":
                     if (data is IByteBuffer dataBufferLT)
@@ -260,10 +260,10 @@ namespace FellowOakDicom.Serialization
                     item = new DicomOtherVeryLong(tag, (IByteBuffer)data);
                     break;
                 case "PN":
-                    item = new DicomPersonName(tag, (string[])data);
+                    item = new DicomPersonName(tag, _jsonTextEncoding, (string[])data);
                     break;
                 case "SH":
-                    item = new DicomShortString(tag, (string[])data);
+                    item = new DicomShortString(tag, _jsonTextEncoding, (string[])data);
                     break;
                 case "SL":
                     if (data is IByteBuffer dataBufferSL)